% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's algorithm returning the shortest distance from a starting vertex of the graph to every other vertex}
\source{

}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{Should be a data.frame with three variables, v1,v2 and w.}

\item{init_node}{Start node.}
}
\description{
Dijkstra's algorithm returning the shortest distance from a starting vertex of the graph to every other vertex
}
