\name{euclidean}
\alias{euclidean}
Euclidean function.
\title{
Euclidean algorithm}
\description{
A function to calculate greatest common divisor of two numbers (x and y) using the Euclidean algorithm based on psudocode from Wikipedia
}
\usage{
euclidean(x, y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Integer
}
  \item{y}{
Integer}
}
\details{
Arguments must be integers.
}
\value{
Greatest common divisor.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Lech Kaczmarczyk and Elis Carlberg Larsson
}
\note{
No notes
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
euclidean(100,1000)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
