% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{A data set containing a graph taken from the Wikipedia page of Dijkstra's algorithm.}
\format{
A data frame with 18 rows and 3 variables.
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
A data set containing a graph taken from the Wikipedia page of Dijkstra's algorithm.
}
\keyword{datasets}
