% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Function to calculate greatest common divisor of two numbers (x and y) using the Euclidean algorithm based on pseudocode from Wikipedia}
\source{
\url{https://en.wikipedia.org/wiki/Euclidean_distance}
}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{a numeric scalars or integers}

\item{y}{a numeric scalars or integers}
}
\description{
Function to calculate greatest common divisor of two numbers (x and y) using the Euclidean algorithm based on pseudocode from Wikipedia
}
